Don't assume traits used as type are trait objs
