+++ tests/ui/rust-2021/ice-unsized-fn-params-2.stderr
@@ -1,18 +1,18 @@
 error[E0782]: expected a type, found a trait
-  --> $DIR/avoid-ice-on-warning-2.rs:4:13
+  --> $DIR/ice-unsized-fn-params-2.rs:7:14
    |
-LL | fn id<F>(f: Copy) -> usize {
-   |             ^^^^
+LL | fn guard(_s: Copy) -> bool {
+   |              ^^^^
    |
    = note: `Copy` it is dyn-incompatible, so it can't be `dyn`
 help: use a new generic type parameter, constrained by `Copy`
    |
-LL | fn id<F, T: Copy>(f: T) -> usize {
-   |        +++++++++     ~
+LL | fn guard<T: Copy>(_s: T) -> bool {
+   |         +++++++++     ~
 help: you can also use an opaque type, but users won't be able to specify the type parameter when calling the `fn`, having to rely exclusively on type inference
    |
-LL | fn id<F>(f: impl Copy) -> usize {
-   |             ++++
+LL | fn guard(_s: impl Copy) -> bool {
+   |              ++++
 
 error: aborting due to 1 previous error
